Extract most logic of the EncryptedSerialized converter into a base class that does not assume specifics about passed in config.
