+++ gobblin-modules/gobblin-crypto/build.gradle
@@ -25,6 +25,7 @@ dependencies {
    */
   compile project(":gobblin-api")
   compile project(":gobblin-modules:gobblin-codecs")
+  compile project(":gobblin-modules:gobblin-metadata")
   compile externalDependency.slf4j
 
   testCompile project(":gobblin-core-base") // for EncryptionConfigParser
